Fixing Achewood and labelling as "random"
Chris Onstad's actually been posting new Achewoods exclusively at
Patreon for a little while now, and the webpage itself switched
over to displaying a random comic on the homepage (or, more
specifically, redirecting to a random comic's page, when loading
just the homepage).  I'm not going to be trying to hook into
Patreon posts, but it'd be fun to start seeing a random Achewood
every day, so I'm also marking this as no longer "on hold."  Fun!